re add validation to model
